+++ truffle/src/com.oracle.truffle.api.debug/src/com/oracle/truffle/api/debug/SteppingStrategy.java
@@ -167,7 +167,7 @@ abstract class SteppingStrategy {
         return new Unwind(depth);
     }
 
-    static SteppingStrategy createUnwind(int depth, Object returnValue) {
+    static SteppingStrategy createUnwind(int depth, DebugValue returnValue) {
         return new Unwind(depth, returnValue);
     }
 
@@ -535,7 +535,7 @@ abstract class SteppingStrategy {
     static final class Unwind extends SteppingStrategy {
 
         private final int depth; // Negative depth
-        private final Object returnValue;
+        private final DebugValue returnValue;
         private int stackCounter;
         ThreadDeath unwind;
 
@@ -543,7 +543,7 @@ abstract class SteppingStrategy {
             this(depth, null);
         }
 
-        Unwind(int depth, Object returnValue) {
+        Unwind(int depth, DebugValue returnValue) {
             this.depth = -depth;
             this.returnValue = returnValue;
         }
@@ -567,7 +567,7 @@ abstract class SteppingStrategy {
         @Override
         Object notifyOnUnwind() {
             if (depth == stackCounter) {
-                return returnValue != null ? returnValue : ProbeNode.UNWIND_ACTION_REENTER;
+                return returnValue != null ? returnValue.get() : ProbeNode.UNWIND_ACTION_REENTER;
             } else {
                 return null;
             }