+++ truffle/src/com.oracle.truffle.api.debug/src/com/oracle/truffle/api/debug/SuspendedEvent.java
@@ -681,11 +681,11 @@ public final class SuspendedEvent {
      * event}.
      *
      * @param frame the frame to unwind
-     * @param immediateReturnValue an Interop value to return
+     * @param immediateReturnValue the value to return
      * @throws IllegalArgumentException when the frame is not on the execution stack of this event
      * @since 21.1.0
      */
-    public void prepareUnwindFrame(DebugStackFrame frame, Object immediateReturnValue) throws IllegalArgumentException {
+    public void prepareUnwindFrame(DebugStackFrame frame, DebugValue immediateReturnValue) throws IllegalArgumentException {
         if (frame.event != this) {
             throw new IllegalArgumentException("The stack frame is not in the scope of this event.");
         }