+++ truffle/CHANGELOG.md
@@ -21,6 +21,7 @@ This changelog summarizes major changes between Truffle versions relevant to lan
 * Added `Shape.getLayoutClass()` as a replacement for `Shape.getLayout().getType()`. Returns the DynamicObject subclass provided to `Shape.Builder.layout`.
 * Changed the default value of `--engine.MultiTier` from `false` to `true`. This should significantly improve the warmup time of Truffle interpreters.
 * Added `SuspendedEvent.prepareUnwindFrame(DebugStackFrame frame, Object forceEarlyReturnValue)` to support forced early return values from a debugger.
+* Added `DebugScope.convertRawValue()` to enable wrapping a raw guest language object into a DebugValue.
 
 ## Version 21.0.0
 * If an `AbstractTruffleException` is thrown from the `ContextLocalFactory`, `ContextThreadLocalFactory` or event listener, which is called during the context enter, the excepion interop messages are executed without a context being entered. The event listeners called during the context enter are: